requirements.txt
django
django-cleanup
Boto3
Pillow
requests
beautifulsoup4
coverage
djangorestframework
django-cors-headers
imdbpy
spotipy
opencv-python
pandas
numpy
tqdm
unidecode
google-api-python-client
wikipedia
python-dotenv